refactor: extract secret.go into a separate pkg
